requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
Django==4.2.3
django-cors-headers==4.1.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.6
PyJWT==1.7.1
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
